pyproject.toml
[tool.poetry]
name = "ratetask"
version = "0.1.0"
description = "Xeneta ratetask"
authors = ["Ramez Salman <ramezsalman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
